+++ rollup.config.js
@@ -1,7 +1,7 @@
 import path from 'path'
+import json from '@rollup/plugin-json'
 import builtinModules from 'builtin-modules'
 import commonjs from 'rollup-plugin-commonjs'
-import json from 'rollup-plugin-json'
 import nodeResolve from 'rollup-plugin-node-resolve'
 import { terser } from 'rollup-plugin-terser'
 import typescript from 'rollup-plugin-typescript'