+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Add documentation about how to enable HTML into README ([#98](https://github.com/marp-team/marp-vscode/issues/98), [#99](https://github.com/marp-team/marp-vscode/pull/99) by [@eric-burel](https://github.com/eric-burel))
 
+### Changed
+
+- Upgrade dependent packages to the latest version ([#102](https://github.com/marp-team/marp-vscode/pull/102))
+
 ## v0.11.0 - 2019-11-07
 
 ### Breaking